included unit tests for commcare and fixed related bugs
